+++ src/graphics/transformation.rs
@@ -1,6 +1,7 @@
-use nalgebra;
+use nalgebra::{Matrix3, Matrix4, Projective2, Vector3, U3, U4};
 use std::ops::Mul;
 
+use crate::graphics::point::Point;
 use crate::graphics::vector::Vector;
 
 /// A 2D transformation matrix.
@@ -9,12 +10,12 @@ use crate::graphics::vector::Vector;
 ///
 /// [`Target`]: struct.Target.html
 #[derive(Debug, Clone, Copy, PartialEq)]
-pub struct Transformation(nalgebra::Matrix4<f32>);
+pub struct Transformation(Matrix4<f32>);
 
 impl Transformation {
     /// Get the identity transformation.
     pub fn identity() -> Transformation {
-        Transformation(nalgebra::Matrix4::identity())
+        Transformation(Matrix4::identity())
     }
 
     /// Creates an orthographic projection.
@@ -33,35 +34,69 @@ impl Transformation {
     ///
     /// You can use this to pan your camera, for example.
     pub fn translate(translation: Vector) -> Transformation {
-        Transformation(nalgebra::Matrix4::new_translation(
-            &nalgebra::Vector3::new(translation.x, translation.y, 0.0),
-        ))
+        Transformation(Matrix4::new_translation(&Vector3::new(
+            translation.x,
+            translation.y,
+            0.0,
+        )))
     }
 
     /// Creates a uniform scale transformation.
     ///
     /// You can use this to zoom your camera, for example.
     pub fn scale(scale: f32) -> Transformation {
-        Transformation(nalgebra::Matrix4::new_scaling(scale))
+        Transformation(Matrix4::new_scaling(scale))
     }
 
     /// Creates a non-uniform scale transformation.
     ///
     /// It allows you to scale each axis independently. You should rarely need
     /// this.
-    pub fn nonuniform_scale(x: f32, y: f32) -> Transformation {
-        Transformation(nalgebra::Matrix4::new_nonuniform_scaling(
-            &nalgebra::Vector3::new(x, y, 1.0),
-        ))
+    pub fn nonuniform_scale(scale: Vector) -> Transformation {
+        Transformation(Matrix4::new_nonuniform_scaling(&Vector3::new(
+            scale.x, scale.y, 1.0,
+        )))
     }
 
     /// Creates a rotation transformation (in radians).
     ///
     /// You can use this to rotate your camera, for example.
     pub fn rotate(rotation: f32) -> Transformation {
-        Transformation(nalgebra::Matrix4::new_rotation(nalgebra::Vector3::new(
-            0.0, 0.0, rotation,
-        )))
+        Transformation(Matrix4::new_rotation(Vector3::new(0.0, 0.0, rotation)))
+    }
+
+    /// Transforms the given point by this transformation.
+    pub fn transform_point(self, point: Point) -> Point {
+        self.0
+            .fixed_slice::<U3, U3>(0, 0)
+            .to_owned()
+            .transform_point(&point)
+    }
+
+    /// Transforms the given vector by this transformation.
+    pub fn transform_vector(self, vector: Vector) -> Vector {
+        self.0
+            .fixed_slice::<U3, U3>(0, 0)
+            .to_owned()
+            .transform_vector(&vector)
+    }
+
+    /// Transforms the given point by the inverse of this transformation.
+    pub fn inverse_transform_point(self, point: Point) -> Point {
+        self.0.fixed_slice::<U3, U3>(0, 0)
+            .to_owned()
+            .try_inverse()
+            .expect("Transformation matrix should only contain invertible operations")
+            .transform_point(&point)
+    }
+
+    /// Transforms the given vector by the inverse of this transformation.
+    pub fn inverse_transform_vector(self, vector: Vector) -> Vector {
+        self.0.fixed_slice::<U3, U3>(0, 0)
+            .to_owned()
+            .try_inverse()
+            .expect("Transformation matrix should only contain invertible operations")
+            .transform_vector(&vector)
     }
 }
 
@@ -86,5 +121,22 @@ impl From<Transformation> for [f32; 16] {
             t.0[8], t.0[9], t.0[10], t.0[11], t.0[12], t.0[13], t.0[14],
             t.0[15],
         ]
+    }
+}
+
+impl From<Projective2<f32>> for Transformation {
+    fn from(projective2: Projective2<f32>) -> Self {
+        let mut matrix =
+            projective2.to_homogeneous().fixed_resize::<U4, U4>(0.0);
+        matrix[15] = 1.0;
+        Transformation(matrix)
+    }
+}
+
+impl Into<Projective2<f32>> for Transformation {
+    fn into(self) -> Projective2<f32> {
+        Projective2::from_matrix_unchecked(Matrix3::from(
+            self.0.fixed_slice::<U3, U3>(0, 0),
+        ))
     }
 }